Fix logical bug in SQL query: implicit rounding in workload computation

+++ School. Distance Learning/ViewModels/Statistics/IndexViewModel.cs
@@ -38,7 +38,7 @@ namespace School._Distance_Learning.ViewModels.Statistics
             {
                 if (WorkingHoursNumber != 0)
                 {
-                    return RealWorkingHoursNumber / WorkingHoursNumber;
+                    return (double)RealWorkingHoursNumber / (double)WorkingHoursNumber;
                 }
 
                 return 0;
